porquim/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.10.1
pydantic-settings==2.7.0

# HTTP
httpx==0.27.0
python-multipart==0.0.12

# IA
openai==1.61.0

# PostgreSQL
asyncpg==0.30.0
